refactor(node): split the 857-line index.ts into focused modules
One file held the MCP wiring, the OPC UA client/session lifecycle, the capability
probes, and all nine tool implementations. Now:

  config.ts      4 lines   environment
  dates.ts      13 lines   toDate
  contract.ts   18 lines   contract + version, staged into build/
  index.ts      95 lines   MCP wiring, signal handling, entry point
  connection.ts 114 lines  client/session lifecycle + capability probes
  tools.ts      676 lines  tool implementations, listTools, callTool dispatch

Tool bodies were moved verbatim: OpcuaTools exposes the same `this.session`,
`this.ensureConnection()` and `this.serverCapabilitiesAggregateFunctions()`
surface by delegating to OpcuaConnection, so not a line of the ~530 lines of tool
logic was retyped. Modules were carved by exact anchors and round-trip checked
against the original text.

Deliberate deviation from the plan: this stops at one tools.ts rather than
tools/*.ts. What made the file hard to work in was the mix of concerns, and that
is now separated — adding a tool touches tools.ts and the contract, nothing else.
Splitting the tool bodies further is cosmetic and was not worth more churn here.

Also broadens the version-literal guard to scan every src/*.ts rather than only
index.ts, since a literal could now reappear in any module. Mutation-tested.

Suite: 69 passed, 1 skipped, 7 smoke, 9 Node unit. typecheck/prettier/ruff clean.
